[RENAME] secure-biometric-device to secure-biometric-interface-integrator
