+++ secure-biometric-interface-integrator/assets/locales/ar/ar.json
