+++ secure-biometric-interface-integrator/example/html/index.html
@@ -11,7 +11,7 @@
 <body>
 
     <div style="width: 300px; margin: 20px">
-        <div id="secure-biometric-device"></div>
+        <div id="secure-biometric-interface-integrator"></div>
     </div>
     
     <div style="width: 300px; margin: 20px">
