+++ secure-biometric-interface-integrator/example/react/package.json
@@ -1,12 +1,12 @@
 {
-    "name": "secure-biometric-device-react-test",
+    "name": "secure-biometric-interface-integrator-react-test",
     "version": "0.1.0",
     "private": true,
     "dependencies": {
         "@testing-library/jest-dom": "^5.16.5",
         "@testing-library/react": "^13.4.0",
         "@testing-library/user-event": "^13.5.0",
-        "secure-biometric-device": "file:../../secure-biometric-device-0.1.0.tgz",
+        "secure-biometric-interface-integrator": "file:../../secure-biometric-interface-integrator-0.1.0.tgz",
         "react": "^18.2.0",
         "react-dom": "^18.2.0",
         "react-scripts": "5.0.1",
