+++ secure-biometric-interface-integrator/lib/standardConstant.js
