+++ secure-biometric-interface-integrator/stories/Sbd.stories.js
@@ -117,7 +117,7 @@ export default {
 
 const renderComponent = (args) => {
   const container = document.createElement("div");
-  container.setAttribute("id", "secure-biometric-device");
+  container.setAttribute("id", "secure-biometric-interface-integrator");
   args.container = container;
   return div(
     {
@@ -159,11 +159,11 @@ Primary.parameters = {
   docs: {
     source: {
       code: `
-      <div id="secure-biometric-device"></div>
+      <div id="secure-biometric-interface-integrator"></div>
 
       // in javascript
       SecureBiometricInterface.init({
-        container: document.getElementById("secure-biometric-device"),
+        container: document.getElementById("secure-biometric-interface-integrator"),
         buttonLabel: "scan_and_verify",
         sbiEnv: {
           env: "Staging",
