+++ secure-biometric-interface-integrator/utility/elements.js
