+++ secure-biometric-interface-integrator/utility/i18n.js
