+++ secure-biometric-interface-integrator/utility/index.js
