+++ secure-biometric-interface-integrator/utility/loadingIndicator.js
