+++ secure-biometric-interface-integrator/utility/localstorage.js
